requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1

psycopg2==2.6.1

pytest==2.8.7
testing.common.database==1.1.0
testing.postgresql==1.3.0
